+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * fix formatting `syntax-rules`
 * improve floating point representation and parsing
 * fix bookmark when hovering over boolean values
+* fix quoted cycle in REPL [#313](https://github.com/jcubic/lips/issues/313)
 
 ## 1.0.0-beta.18
 ### Breaking